requirements.txt
python-telegram-bot==21.4
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.8.2
PyYAML==6.0.2
